backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
authlib>=1.3.0
bcrypt>=4.1.0
python-multipart>=0.0.9
redis>=5.0.0
google-cloud-storage>=2.16.0
python-magic>=0.4.27
langchain>=0.2.0
langchain-groq>=0.1.0
pandas>=2.2.0
openpyxl>=3.1.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pydantic-settings>=2.2.0
groq>=0.30.0